Begin extraction of qvm-template functionality into module
